@@ -2,10 +2,10 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>mockserver</ToolCommandName>
-    <Version>2.2.0</Version>
+    <Version>3.0.0</Version>
     <InformationalVersion>This is a release package</InformationalVersion>
     <IsPackable>true</IsPackable>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
@@ -19,8 +19,8 @@
     <PackageTags>Command Line; Standalone; WireMock; Server; Proxy; Mock Http;</PackageTags>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Topshelf" Version="4.2.1.218-develop" />
-    <PackageReference Include="WireMock.Net.StandAlone" Version="1.0.38" />
+    <PackageReference Include="Topshelf" Version="4.3.0" />
+    <PackageReference Include="WireMock.Net.StandAlone" Version="1.5.47" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\TestEssentials.ToolKit.WireMock\TestEssentials.ToolKit.WireMock.csproj" />
